Factor transparent transaction components into a separate module.
