+++ zcash_primitives/src/transaction/components.rs
@@ -1,13 +1,11 @@
 //! Structs representing the components within Zcash transactions.
 
-use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
+use byteorder::{ReadBytesExt, WriteBytesExt};
 
 use std::io::{self, Read, Write};
 
 #[cfg(feature = "zfuture")]
 use std::convert::TryFrom;
-
-use crate::legacy::Script;
 
 #[cfg(feature = "zfuture")]
 use crate::{
@@ -18,111 +16,16 @@ use crate::{
 pub mod amount;
 pub mod sapling;
 pub mod sprout;
+pub mod transparent;
 pub use self::{
     amount::Amount,
     sapling::{OutputDescription, SpendDescription},
     sprout::JSDescription,
+    transparent::{OutPoint, TxIn, TxOut},
 };
 
 // π_A + π_B + π_C
 pub const GROTH_PROOF_SIZE: usize = 48 + 96 + 48;
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct OutPoint {
-    hash: [u8; 32],
-    n: u32,
-}
-
-impl OutPoint {
-    pub fn new(hash: [u8; 32], n: u32) -> Self {
-        OutPoint { hash, n }
-    }
-
-    pub fn read<R: Read>(mut reader: R) -> io::Result<Self> {
-        let mut hash = [0u8; 32];
-        reader.read_exact(&mut hash)?;
-        let n = reader.read_u32::<LittleEndian>()?;
-        Ok(OutPoint { hash, n })
-    }
-
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(&self.hash)?;
-        writer.write_u32::<LittleEndian>(self.n)
-    }
-
-    pub fn n(&self) -> u32 {
-        self.n
-    }
-
-    pub fn hash(&self) -> &[u8; 32] {
-        &self.hash
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct TxIn {
-    pub prevout: OutPoint,
-    pub script_sig: Script,
-    pub sequence: u32,
-}
-
-impl TxIn {
-    #[cfg(feature = "transparent-inputs")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "transparent-inputs")))]
-    pub fn new(prevout: OutPoint) -> Self {
-        TxIn {
-            prevout,
-            script_sig: Script::default(),
-            sequence: std::u32::MAX,
-        }
-    }
-
-    pub fn read<R: Read>(mut reader: &mut R) -> io::Result<Self> {
-        let prevout = OutPoint::read(&mut reader)?;
-        let script_sig = Script::read(&mut reader)?;
-        let sequence = reader.read_u32::<LittleEndian>()?;
-
-        Ok(TxIn {
-            prevout,
-            script_sig,
-            sequence,
-        })
-    }
-
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        self.prevout.write(&mut writer)?;
-        self.script_sig.write(&mut writer)?;
-        writer.write_u32::<LittleEndian>(self.sequence)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct TxOut {
-    pub value: Amount,
-    pub script_pubkey: Script,
-}
-
-impl TxOut {
-    pub fn read<R: Read>(mut reader: &mut R) -> io::Result<Self> {
-        let value = {
-            let mut tmp = [0u8; 8];
-            reader.read_exact(&mut tmp)?;
-            Amount::from_nonnegative_i64_le_bytes(tmp)
-        }
-        .map_err(|_| io::Error::new(io::ErrorKind::InvalidData, "value out of range"))?;
-        let script_pubkey = Script::read(&mut reader)?;
-
-        Ok(TxOut {
-            value,
-            script_pubkey,
-        })
-    }
-
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(&self.value.to_i64_le_bytes())?;
-        self.script_pubkey.write(&mut writer)
-    }
-}
 
 #[cfg(feature = "zfuture")]
 fn to_io_error(_: std::num::TryFromIntError) -> io::Error {
